% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv.sensemakr-package.R
\docType{package}
\name{iv.sensemakr-package}
\alias{iv.sensemakr}
\alias{iv.sensemakr-package}
\title{iv.sensemakr: Sensitivity Analysis Tools for Instrumental Variable Estimates}
\description{
Implements a suite of sensitivity analysis tools for instrumental variable
estimates, as described in Cinelli and Hazlett (2025).
}
\details{
The main functions of the package are \code{\link{iv_fit}}, which fits an IV
regression model using the Anderson-Rubin approach, and
\code{\link{sensemakr.iv_fit}}, which computes sensitivity analysis results
for objects of class \code{\link{iv_fit}}.
}
\references{
Cinelli, C. and Hazlett, C. (2025), "An Omitted Variable Bias Framework for
Sensitivity Analysis of Instrumental Variables." Biometrika.
\doi{10.1093/biomet/asaf004}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/carloscinelli/iv.sensemakr}
  \item Report bugs at \url{https://github.com/carloscinelli/iv.sensemakr/issues}
}

}
\author{
\strong{Maintainer}: Carlos Cinelli \email{carloscinelli@hotmail.com}

Authors:
\itemize{
  \item Chad Hazlett
}

}
\keyword{internal}
