% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakr.R
\docType{package}
\name{iv.sensemakr}
\alias{iv.sensemakr}
\title{iv.sensemakr}
\description{
The \code{iv.sensemakr} package implements a suite of sensitivity analysis tools that makes it easier to
understand the impact of omitted variables in instrumental variable regression, as discussed in Cinelli and Hazlett (2023).
}
\details{
The main functions of the package are \code{\link{iv_fit}}, which fits an IV regression model, using the Anderson-Rubin approach, and \code{\link{sensemakr}}, which computes the most common sensitivity analysis results for the \code{iv_fit} object.

After running \code{iv_fit} or \code{sensemakr} you may directly use the plot, print, coef, and summary methods in the returned object.

More information can be found on the help documentation, vignettes and related papers.
}
\examples{

# loads package
library(iv.sensemakr)

# loads dataset
data("card")

# prepares data
y <- card$lwage  # outcome
d <- card$educ   # treatment
z <- card$nearc4 # instrument
x <- model.matrix( ~ exper + expersq + black + south + smsa + reg661 + reg662 +
                     reg663 + reg664 + reg665+ reg666 + reg667 + reg668 + smsa66,
                   data = card) # covariates
# fits IV model
card.fit <- iv_fit(y,d,z,x)

# see results
card.fit

# runs sensitivity analysis
card.sens <- sensemakr(card.fit, benchmark_covariates = c("black", "smsa"))

# see results
card.sens

# sensitivity contour plot
plot(card.sens, lim = 0.09)
}
\references{
Cinelli, C. and Hazlett, C. (2023), "An Omitted Variable Bias Framework for Sensitivity Analysis of Instrumental Variables."
}
