% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.iv.sensemakr}
\alias{plot.iv.sensemakr}
\title{Sensitivity analysis plots for IV sensemakr}
\usage{
\method{plot}{iv.sensemakr}(x, sensitivity.of = c("ci", "lwr", "upr", "t-value"), parm = "iv", ...)
}
\arguments{
\item{x}{an object of class \code{iv.sensemakr} created with the \code{\link{sensemakr}} function.}

\item{sensitivity.of}{should the contour plot show adjusted lower limits of confidence intervals (\code{"lwr"}), upper limit of confidence intervals (\code{"upr"}) or t-values (\code{"t-value"})?}

\item{parm}{contour plots of which estimate? Options are \code{iv} for instrumental variable estimates, \code{fs} for first-stage estimates, and \code{rf} for reduced-form estimates.}

\item{...}{further arguments and graphical parameters.}
}
\description{
This function provides the contour plots of the sensitivity analysis results obtained with the function \code{\link{sensemakr}} for IV. It is basically a dispatcher to the core plot function \code{\link{ovb_contour_plot}}.
}
