% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_stats.R
\name{robustness_value}
\alias{robustness_value}
\alias{extreme_robustness_value}
\alias{xrv}
\alias{rv}
\alias{extreme_robustness_value.iv_fit}
\alias{robustness_value.iv_fit}
\title{Computes the (extreme) robustness value for IV}
\usage{
robustness_value(...)

extreme_robustness_value(...)

xrv(...)

rv(...)

\method{extreme_robustness_value}{iv_fit}(model, parm = "iv", q = 1, alpha = 0.05, min = T, ...)

\method{robustness_value}{iv_fit}(model, parm = "iv", q = 1, alpha = 0.05, min = T, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}

\item{model}{an \code{\link{iv_fit}} model}

\item{parm}{parameter for which the robustness value is computed. Default is \code{iv}, meaning that the robustness value of the IV estimate is computed. Other options are to compute the robustness value of auxiliary estimates, such as the first stage (\code{fs}) or the reduced form (\code{rf}).}

\item{q}{percent change of the effect estimate that would be deemed problematic. Default is 1, which means a reduction (increase) of 100\% of the current effect estimate (bring estimate to zero). It has to be greater than zero.}

\item{alpha}{significance level.}

\item{min}{in many cases, researchers are interested in biases as large or larger than a certain amount (for instance, the strength of confounding to bring a positive estimate to zero or below). Setting \code{min = TRUE} (default) computes the robustness value for such cases. Setting \code{min = FALSE} computes the robustness value for a bias of exactly \code{q}.}
}
\description{
Description here.
}
