% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakr.R
\name{summary.iv.sensemakr}
\alias{summary.iv.sensemakr}
\alias{print.summary.iv.sensemakr}
\title{Sensitivity analysis print and summary methods for \code{iv.sensemakr}}
\usage{
\method{summary}{iv.sensemakr}(object, ...)

\method{print}{summary.iv.sensemakr}(x, digits = 3, ...)
}
\arguments{
\item{object}{an object of class \code{\link{sensemakr}}.}

\item{...}{arguments passed to other methods.}

\item{x}{an object of class \code{\link{sensemakr}}.}

\item{digits}{minimal number of \emph{significant} digits.}
}
\description{
The \code{print} and \code{summary} methods provide verbal descriptions of the sensitivity analysis results
obtained with the function \code{\link{sensemakr}}.
}
