% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_functions.R
\name{summary.iv_fit}
\alias{summary.iv_fit}
\alias{print.iv_fit}
\title{print and summary methods for \code{iv_fit}}
\usage{
\method{summary}{iv_fit}(object, ...)

\method{print}{iv_fit}(x, digits = 3, ...)
}
\arguments{
\item{object}{an object of class \code{\link{iv_fit}}.}

\item{...}{arguments passed to other methods}

\item{x}{an object of class \code{\link{iv_fit}}.}

\item{digits}{minimal number of significant digits}
}
\description{
The print and summary methods provide verbal descriptions of the results obtained with the function \code{\link{iv_fit}}.
}
