% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_stats.R
\name{sensitivity_stats}
\alias{sensitivity_stats}
\alias{sensitivity_stats.iv_fit}
\title{Sensitivity statistics for instrumental variable estimates}
\usage{
sensitivity_stats(...)

\method{sensitivity_stats}{iv_fit}(model, parm = "iv", q = 1, alpha = 0.05, min = TRUE, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}

\item{model}{a model created with the function \code{\link{iv_fit}}.}

\item{parm}{contour plots of which estimate? Options are \code{iv} for instrumental variable estimates, \code{fs} for first-stage estimates, and \code{rf} for reduced-form estimates.}

\item{q}{percent change of the effect estimate that would be deemed problematic. Default is 1, which means a reduction of 100\% of the current effect estimate (bring estimate to zero).}

\item{alpha}{significance level.}

\item{min}{should we consider biases as large or larger than a certain amount? Default is \code{TRUE}.}
}
\description{
Convenience function that computes robustness values for IV estimates as well as auxiliary first stage and reduced form regressions.
}
